requirements.txt
asgiref==3.2.3
Django==3.0.2
djangorestframework==3.11.0
imutils==0.5.3
numpy==1.18.1
opencv-python==4.1.2.30
Pillow==7.0.0
pytesseract==0.3.1
pytz==2019.3
sqlparse==0.3.0
